Add files via upload
The file "Speech_Emotion_Recognition.py" holds the main logic of this project, the multi-input deep learning model. Along with the model, the code also contains a custom augmentation for the spectogram images but since we are applying CutMix as well, for now the augmentation is set to False for all the images. In future, we plan on combining both, the custom augmentation and CutMix, such that they don't overwhelm the model, and add up positively.